refactor(script): split dialog and PC buildins out of script.c
Issue #61 fifth step: move dialog/UI helpers into buildin_dialog.c and PC/status helpers into buildin_pc.c.
