+++ vcproj/map-server.vcxproj
@@ -211,6 +211,8 @@
     <ClCompile Include="..\src\map\buildin_map.c" />
     <ClCompile Include="..\src\map\buildin_npc.c" />
     <ClCompile Include="..\src\map\buildin_party.c" />
+    <ClCompile Include="..\src\map\buildin_dialog.c" />
+    <ClCompile Include="..\src\map\buildin_pc.c" />
     <ClCompile Include="..\src\map\skill.c" />
     <ClCompile Include="..\src\map\skill_castend_damage.c" />
     <ClCompile Include="..\src\map\skill_castend_ground.c" />
